Got depth working in sim.
Lowkey don't need to do translation of RBG camera to depth camera in sim. Their frames are a pure rotation.

Idk that this will be true on the real bot, but if so, it still does the mapping procedure. It just has a guard so pure rotations (which dont require image warping) dont break stuff